main algorithm

+++ Ex2/ex2/queue_and_package.c
@@ -279,7 +279,7 @@ void PrintQueues(QUEUE* head) {
 	}
 }
 
-int SumActiveLinksWeights(QUEUE* head) {
+float SumActiveLinksWeights(QUEUE* head) {
 	QUEUE* queue = head;
 	float sum = 0;
